Рефакторинг кода модулей
