+++ index.html
@@ -234,6 +234,5 @@
       </section>
     </template>
     <script src="js/main.js" type="module"></script>
-    <script src="js/render_preview.js" type="module"></script>
 </body>
 </html>
