+++ js/main.js
@@ -1,1 +1,4 @@
-import { createPhotos } from './data.js';
+import { renderPreview } from './images.js';
+
+const PREVIEWS_COUNT = 25;
+renderPreview(PREVIEWS_COUNT);
